+++ mirc.h
@@ -37,6 +37,6 @@
 #define MIRC_LIGHTCYAN     11
 #define MIRC_WHITE          0
 
-const uint32_t mirc_colors[16];
+const uint8_t mirc_colors[16];
 
 #endif
